feat: export check entry function
